JG # Add code in bloc for add item

+++ lib/src/pro/ui/sales/sales_order/bloc/sales_order_event.dart
@@ -9,7 +9,7 @@ abstract class SalesOrderEvent extends Equatable {
 }
 
 class AddItemButtonPress extends SalesOrderEvent {
-  final int setQty;
+  final double setQty;
   final String searchText;
   final String itemNumber;
   const AddItemButtonPress(
